requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
sqlparse==0.5.1
whitenoise==6.7.0
